# Implement a tire model and compute the normal force