--- a/docs/tutorials/four_bar_linkage.ipynb
+++ b/docs/tutorials/four_bar_linkage.ipynb
@@ -722,7 +722,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.9.18"
+   "version": "3.11.7"
   }
  },
  "nbformat": 4,