--- a/docs/tutorials/my_first_bicycle.ipynb
+++ b/docs/tutorials/my_first_bicycle.ipynb
@@ -99,6 +99,16 @@
    "metadata": {},
    "outputs": [],
    "source": [
+    "bicycle_def.rear_tire.compute_normal_force = True\n",
+    "bicycle_def.front_tire.compute_normal_force = True"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
     "# Manually run all steps\n",
     "bicycle_def.define_connections()\n",
     "bicycle_def.define_objects()\n",
@@ -204,7 +214,9 @@
   },
   {
    "cell_type": "markdown",
-   "metadata": {},
+   "metadata": {
+    "jp-MarkdownHeadingCollapsed": true
+   },
    "source": [
     "Congratulations you've just created your first bicycle model using BRiM!\n",
     "\n",
@@ -634,6 +646,15 @@
    ]
   },
   {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "uaux_zero = dict.fromkeys(system_def.u_aux.diff(me.dynamicsymbols._t).col_join(system_def.u_aux), 0)"
+   ]
+  },
+  {
    "cell_type": "markdown",
    "metadata": {
     "jupyter": {
@@ -938,7 +959,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.9.18"
+   "version": "3.11.7"
   }
  },
  "nbformat": 4,