--- a/docs/tutorials/my_first_bicycle.ipynb
+++ b/docs/tutorials/my_first_bicycle.ipynb
@@ -99,16 +99,6 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "bicycle_def.rear_tire.compute_normal_force = True\n",
-    "bicycle_def.front_tire.compute_normal_force = True"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
     "# Manually run all steps\n",
     "bicycle_def.define_connections()\n",
     "bicycle_def.define_objects()\n",
@@ -214,9 +204,7 @@
   },
   {
    "cell_type": "markdown",
-   "metadata": {
-    "jp-MarkdownHeadingCollapsed": true
-   },
+   "metadata": {},
    "source": [
     "Congratulations you've just created your first bicycle model using BRiM!\n",
     "\n",
@@ -646,15 +634,6 @@
    ]
   },
   {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "uaux_zero = dict.fromkeys(system_def.u_aux.diff(me.dynamicsymbols._t).col_join(system_def.u_aux), 0)"
-   ]
-  },
-  {
    "cell_type": "markdown",
    "metadata": {
     "jupyter": {
@@ -959,7 +938,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.11.7"
+   "version": "3.11.8"
   }
  },
  "nbformat": 4,