--- a/src/brim/bicycle/tires.py
+++ b/src/brim/bicycle/tires.py
@@ -4,7 +4,15 @@ from __future__ import annotations
 from typing import Any
 
 from sympy import MutableDenseMatrix as Matrix
-from sympy.physics.mechanics import Point, System, Vector, cross, dynamicsymbols
+from sympy.physics.mechanics import (
+    Force,
+    Point,
+    System,
+    Torque,
+    Vector,
+    cross,
+    dynamicsymbols,
+)
 
 from brim.bicycle.grounds import FlatGround, GroundBase
 from brim.bicycle.wheels import KnifeEdgeWheel, ToroidalWheel, WheelBase
@@ -155,8 +163,29 @@ class TireBase(ConnectionBase):
         self._on_ground = bool(value)
 
 
-class NonHolonomicTire(TireBase):
-    """Tire model connection based on non-holonomic constraints."""
+class InContactTire(TireBase):
+    """Generic tire model that is in contact with the ground.
+
+    Explanation
+    -----------
+    This tire model is a generic tire model that is defined to be in contact with the
+    ground. It is used as a base class for other tire models that are in contact with
+    the ground.
+
+    Attributes
+    ----------
+    compute_normal_force : bool
+        Flag to indicate if the normal force of the tire model should be computed.
+        Default is True.
+    no_longitudinal_slip : bool
+        Flag to indicate if the tire model has no longitudinal slip. If True, a
+        nonholonomic constraint is added to the system to enforce no slip in the
+        longitudinal direction. Default is False.
+    no_lateral_slip : bool
+        Flag to indicate if the tire model has no lateral slip. If True, a nonholonomic
+        constraint is added to the system to enforce no slip in the lateral direction.
+        Default is False.
+    """
 
     required_models: tuple[ModelRequirement, ...] = (
         ModelRequirement("ground", FlatGround, "Submodel of the ground."),
@@ -168,40 +197,54 @@ class NonHolonomicTire(TireBase):
 
     def __init__(self, name: str) -> None:
         super().__init__(name)
-        self._compute_normal_force = False
-
-    @property
-    def compute_normal_force(self) -> bool:
-        """Boolean whether the normal force should be computed."""
-        return self._compute_normal_force
-
-    @compute_normal_force.setter
-    def compute_normal_force(self, value: bool) -> None:
-        self._compute_normal_force = bool(value)
+        self.compute_normal_force: bool = True
+        self.no_longitudinal_slip: bool = False
+        self.no_lateral_slip: bool = False
 
     @property
     def descriptions(self) -> dict[Any, str]:
-        """Dictionary of descriptions of the non-holonomic tire model's attributes."""
+        """Dictionary of descriptions of the lateral tire model's attributes."""
         descriptions = super().descriptions
         if self.compute_normal_force:
-            descriptions.update({
-                self.u[0]: "Auxiliary generalized speed to determine the normal force.",
-                self.symbols["Fz"]: "Negative of the normal force of the tire model.",
-            })
+            descriptions[self.u[0]] = (
+                f"Auxiliary generalized speed to determine the normal force of "
+                f"'{self.name}'.")
+        if "Fx" in self.symbols:
+            descriptions["Fx"] = f"Longitudinal force of tire model '{self.name}'."
+        if "Fy" in self.symbols:
+            descriptions["Fy"] = f"Lateral force of tire model '{self.name}'."
+        if "Fz" in self.symbols:
+            descriptions["Fz"] = f"Normal force of tire model '{self.name}'."
+        if "Mx" in self.symbols:
+            descriptions["Mx"] = (
+                f"Rolling resistance moment of tire model '{self.name}'.")
+        if "Mz" in self.symbols:
+            descriptions["Mz"] = f"Self aligning moment of tire model '{self.name}'."
         return descriptions
 
     def _define_objects(self) -> None:
         """Define the objects of the tire model."""
         super()._define_objects()
+        syms_to_add = ["Fx", "Fy", "Fz", "Mx", "Mz"]
         if self.compute_normal_force:
             self.u = Matrix([dynamicsymbols(self._add_prefix("uaux_z"))])
-            self.symbols["Fz"] = dynamicsymbols(self._add_prefix("Fz"))
+        if self.no_longitudinal_slip:
+            syms_to_add.remove("Fx")
+        if self.no_lateral_slip:
+            syms_to_add.remove("Fy")
+            syms_to_add.remove("Mz")
+        if self.no_lateral_slip and self.no_longitudinal_slip:
+            syms_to_add.remove("Mx")
+        self.symbols.update({
+            name: dynamicsymbols(self._add_prefix(name)) for name in syms_to_add
+        })
 
     def _define_kinematics(self) -> None:
         """Define the kinematics of the tire model."""
         super()._define_kinematics()
         self._set_pos_contact_point()
-        if self.ground.origin in self.contact_point._pos_dict:
+        if (self.no_longitudinal_slip and self.no_lateral_slip
+            and self.ground.origin in self.contact_point._pos_dict):
             self.contact_point.set_vel(self.ground.frame, 0)
             self.wheel.center.set_vel(
                 self.ground.frame,
@@ -215,25 +258,57 @@ class NonHolonomicTire(TireBase):
             self.auxiliary_handler.add_noncontributing_force(
                 self.contact_point, direction, self.u[0], self.symbols["Fz"])
 
+    def _define_loads(self) -> None:
+        """Define the loads of the tire model."""
+        super()._define_loads()
+        tire_force = (
+            self.symbols.get("Fx", 0) * self.longitudinal_axis +
+            self.symbols.get("Fy", 0) * self.lateral_axis
+        )
+        if tire_force != 0:
+            self.system.add_loads(Force(self.contact_point, tire_force))
+        tire_torque = (
+            self.symbols.get("Mx", 0) * self.longitudinal_axis +
+            self.symbols.get("Mz", 0) * self.ground.get_normal(self.contact_point)
+        )
+        if tire_torque != 0:
+            self.system.add_loads(Torque(self.wheel.frame, tire_torque))
+
     def _define_constraints(self) -> None:
         """Define the constraints of the tire model."""
         super()._define_constraints()
-        aux_vel_cp = self.auxiliary_handler.get_auxiliary_velocity(self.contact_point)
-        aux_vel_wc = self.auxiliary_handler.get_auxiliary_velocity(self.wheel.center)
-        aux_vel_gnd = self.auxiliary_handler.get_auxiliary_velocity(self.ground.origin)
-        normal = self.ground.get_normal(self.contact_point)
-        tangent_vectors = self.ground.get_tangent_vectors(self.contact_point)
-        v0 = (self.wheel.center.pos_from(self.ground.origin).dt(self.ground.frame)
-              + cross(self.wheel.frame.ang_vel_in(self.ground.frame),
-                      self.contact_point.pos_from(self.wheel.center)))
-        aux_v0 = aux_vel_wc - aux_vel_gnd + aux_vel_cp
-        self.system.add_nonholonomic_constraints(
-            v0.dot(tangent_vectors[0]) + aux_v0.dot(tangent_vectors[0]),
-            v0.dot(tangent_vectors[1]) + aux_v0.dot(tangent_vectors[1]))
+        if self.no_longitudinal_slip and self.no_lateral_slip:
+            no_slip_vectors = self.ground.get_tangent_vectors(self.contact_point)
+        elif self.no_longitudinal_slip:
+            no_slip_vectors = (self.longitudinal_axis,)
+        elif self.no_lateral_slip:
+            no_slip_vectors = (self.lateral_axis,)
+        aux_cp = self.auxiliary_handler.get_auxiliary_velocity(self.contact_point)
+        if self.no_longitudinal_slip or self.no_lateral_slip:
+            aux_wc = self.auxiliary_handler.get_auxiliary_velocity(self.wheel.center)
+            aux_gnd = self.auxiliary_handler.get_auxiliary_velocity(self.ground.origin)
+            v0 = (self.wheel.center.pos_from(self.ground.origin).dt(self.ground.frame)
+                + cross(self.wheel.frame.ang_vel_in(self.ground.frame),
+                        self.contact_point.pos_from(self.wheel.center)))
+            aux_v0 = aux_wc - aux_gnd + aux_cp
+            for vector in no_slip_vectors:
+                self.system.add_nonholonomic_constraints(
+                    v0.dot(vector) + aux_v0.dot(vector))
         if not self.on_ground:
+            normal = self.ground.get_normal(self.contact_point)
             self.system.add_holonomic_constraints(
                 self.contact_point.pos_from(self.ground.origin).dot(normal))
             self.system.velocity_constraints = [
                 self.system.holonomic_constraints[0].diff(dynamicsymbols._t) +
-                aux_vel_cp.dot(normal), *self.system.nonholonomic_constraints
+                aux_cp.dot(normal), *self.system.nonholonomic_constraints
             ]
+
+
+class NonHolonomicTire(InContactTire):
+    """Tire model connection based on nonholonomic constraints."""
+
+    def __init__(self, name: str) -> None:
+        super().__init__(name)
+        self.compute_normal_force = False
+        self.no_lateral_slip = True
+        self.no_longitudinal_slip = True