… .github/workflows/tests.yaml
… .pre-commit-config.yaml
… docs/guides/component_implementation.rst
… poetry.lock
… pyproject.toml
… src/brim/bicycle/whipple_bicycle.py
… src/brim/core/__init__.py
… src/brim/core/auxiliary.py
… src/brim/core/base_classes.py
… src/brim/core/singleton.py
… src/brim/other/rolling_disc.py
… src/brim/utilities/testing.py
… src/brim/utilities/utilities.py
… tests/bicycle/test_tire_models.py
… tests/bicycle/test_tires.py
… tests/bicycle/test_whipple_bicycle.py
… tests/core/test_auxiliary.py
… tests/core/test_base_classes.py
… tests/rider/test_hip_joints.py
… tests/rider/test_rider.py