This PR aims to implement a basic setup for creating more advanced tire models. The most difficult issue to solve here is to reliably compute the normal force by introducing auxiliary speeds and adding noncontributing loads to the system.

The main problem of these auxiliary speeds is to ensure correct propagation throughout the velocity graph and to take them into account in the velocity constraints. Related is the introduction of the `velocity_constraints` attribute in [sympy#26374](https://github.com/sympy/sympy/pull/26374).

I have designed a new utility called `AuxiliaryDataHandler` to tackle this issue. The basic idea is to keep all auxiliary speeds separate and introduce them at the end of the define kinematics step. The position graph of points and the orientation graph of reference frames do not contain any cycles. Therefore, you can convert them into a tree representation with the inertial frame as the root of the orientation graph and a fixed point in the inertial frame as the root of the position graph. The advantage of this tree representation is that if you know that an auxiliary speed should be applied to a certain point, then you can easily determine to what points this speed should propagate (namely its children). The auxiliary speed of each point can is also remembered such that they can be requested when determining the velocity constraints.
The load objects of the noncontributing loads are defined as disconnected points with the auxiliary velocity in the inertial frame. It is known that they do no work, so all other terms (if there are any) must fall out anyway.

The main disadvantage in the current setup is that this auxiliary data handler must be shared among all BRiM objects. This does break a bit with the idea that every model is its own system, but I do not really see another way as auxiliary speeds from other systems will have to be able to propagate throughout everything. Some consequences of this decision involve the introduction of the `ModelBase.is_root` attribute and that the noncontributing loads/auxiliary speeds are only added to the root system.

Another fun feature of the `AuxiliaryDataHandler` is its `_compute_velocity` method, which tries to compute the velocity of a point based on `v2pt_theory` or `v1pt_theory` before falling back to the time derivative of the position vector. I am not entirely sure if I'll keep this feature or will remove its usage. This should be decided based on some benchmarks.

TODO:
- [x] Correctly compute the normal force of both the rear and front tire.
- [x] Implement some basic testing.
- [x] Compare benchmarks of different methods to compute the velocity of points.
- [x] Fix failing tests.
- [x] Design a generic tire model interface.
- [x] Ensure full line coverage with the tests.
- [x] Add an explanation section to the `AuxiliaryDataHandler`.
- [x] Add a disclaimer that one should aim to compute velocities based on other points according to the tree structure of the position tree. It is kind of something you never encounter to violate anyways, but it is one of the requirements of the auxiliary data handler to function correctly.
- [x] Validate tire model.

Future features could be:
- Support for noncontributing torques.
- More advanced and specific tire models.
- Make a tutorial on how to work with noncontributing forces. Users will have to know how to solve and use the auxiliary equations after forming the equations of motion with Kane's method.