@moorepants 2024-03-22
Will this implement the functionality to be able to make a tire model as well as the tire model? For the later, there are some considerations to make, like what conventions to use, whether it is a linear or nonlinear model, what model it is, etc.

For a full nonlinear tire model the forces all depend on the normal force, but if that is a noncontributing force for the basic formulation we have. In very general tire force models, the lateral forces and various moments can be nonlinear in the normal force which means you can't simply solve for the noncontributing force algebraically, you'd need to have an iterative numeric solver to handle that. I think that some tire models implement the vertical tire compliance to get around this issue. Then the normal force is a function of the new state variables and there is no need to try to solve for this normal force.