@tjstienstra 2024-03-22
> Will this implement the functionality to be able to make a tire model as well as the tire model?

The main aim is the functionality to be able to make a tire model. Implementing a tire model is more like a second.

I specifically had in mind to first implement a more generic tire model where the tire forces and torques are just dynamicsymbols. I would expect it to result in something like the following:
```python
class InContactTireTireBase):
    """Generic tire model that is in contact with the ground."""

    def __init__(self, name: str) -> None:
        super().__init__(name)
        self.compute_normal_force: bool = False
        self.no_longitudinal_slip: bool = False
        self.no_lateral_slip: bool = False

    def _define_objects(self) -> None:
        """Define the objects of the tire model."""
        super()._define_objects()
        self.symbols.update({
            name: dynamicsymbols(self._add_prefix(name))
            for name in ("Fx", "Fy", "Fz", "Mx", "Mz")
        })
        if self.compute_normal_force:
            self.u = Matrix([dynamicsymbols(self._add_prefix("uaux_z"))])
        if self.no_longitudinal_slip:
            self.symbols["Fx"] = S.Zero
        if self.no_lateral_slip:
            self.symbols["Fy"] = S.Zero
            self.symbols["Mz"] = S.Zero
        if self.no_lateral_slip and self.no_longitudinal_slip:
            for name in ("Fx", "Fy", "Mx", "Mz"):
                self.symbols[name] = S.Zero

    def _define_kinematics(self) -> None:
        super()._define_kinematics()
        self._set_pos_contact_point()
        super()._define_kinematics()
        self._set_pos_contact_point()
        if self.compute_normal_force:
            direction = -self.ground.get_normal(self.contact_point)
            if self.on_ground:
                direction = -direction
            self.auxiliary_handler.add_noncontributing_force(
                self.contact_point, direction, self.u[0], self.symbols["Fz"])

    def _define_loads(self) -> None:
        super()._define_loads()
        # TODO add the loads based on the tire model.

    def _define_constraints(self) -> None:
        super()._define_constraints()
        if self.no_longitudinal_slip and self.no_lateral_slip:
            # TODO add both nonholonomic constraints for no slip.
            pass
        elif self.no_longitudinal_slip:
            # TODO compute nonholonomic constraint for no longitudinal slip.
            pass
        elif self.no_lateral_slip:
            # TODO compute nonholonomic constraint for no lateral slip.
            pass
        if not self.on_ground:
            # TODO add the holonomic constraints.
            pass
```
The nice thing about the above setup is that this class can be inherited and one can simply specify `self.symbols["Fx] = some_function(self.symbols["Fz"])`. Here one can choose for more details, the only constrain is that no compliance is modeled.

About implementing a linear or nonlinear model. I don't expect to implement full nonlinear models right now as that would result in a lot of complexity as you mention. A model I will probably start with is the same one used in the kickplate model.