@moorepants 2024-03-22
> A model I will probably start with is the same one used in the kickplate model.

Ok.