@tjstienstra 2024-03-22
Benchmark results of the Whipple bicycle following Moore's convention:
|                                                                                   | Time | #operation before CSE | #operations after CSE |
|-----------------------------------------------------------------------------------|------|-----------------------|-----------------------|
| main 2ed1713f737d4ec1dc1203cb3d558a27f5d81552                                     | 3.04 | 352210                | 2291                  |
| this PR with custom velocity computation da346de773e073ec912afeab853b16db68c661cf | 3.49 | 368849                | 2762                  |
| this PR with Point.vel                                                            | 2.85 | 266654                | 2526                  |
| this PR with custom velocity computation and ICR assumption front wheel           | 3.34 | 388994                | 2484                  |
| this PR with Point.vel and ICR assumption front wheel                             | 3.10 | 352210                | 2291                  |

ICR assumption front wheel refers to whether the front wheel center's velocity is computed based on the assumption that the contact point is an instantaneous center of rotation. The problem with wanting to apply this assumption is that the velocity tree doesn't match the shape of the position tree resulting in the addition of the wrong auxiliary velocities. I will start thinking about how to resolve this problem, but it will require more active tracking of the graphs and that users/modelers notify the auxiliary data handler when they apply these assumptions.

At least for now, I can conclude that it is best to use `Point.vel` as taking time derivatives of position vectors seems to give really good results.