@tjstienstra 2024-03-23
The benchmark results when using v2pt_theory above just differentiating the vector have always surprised me, as you would normally expect that it is cheaper to use such a neat rule instead of expressing a vector into a frame and then differentiating it.
Therefore, I had a bit of a deeper dive into this issue. The problem seems to be a bad choice of how we apply our cross-product. In v2pt_theory we apply our cross-product using the unit vectors of the frames used in the position vector, while it would be way more efficient to use the unit vectors of the angular velocity. I just experimented with this a little by changing `self.set_vel(outframe, v + (omega ^ dist))` in `v2pt_theory` to `self.set_vel(outframe, v - (dist ^ omega))`. This change resulted in 2.59 seconds, 232020 operations before CSE and 1919 operations after CSE.