@moorepants 2024-03-23
I think that Peter may have come across this before, also showing that equations were longer when using a method that you would think would be shorter.