@tjstienstra 2024-03-23
> I think that Peter may have come across this before, also showing that equations were longer when using a method that you would think would be shorter.

Yeah, wouldn't surprise me.

I may open an issue in SymPy on whether we shall make this change. I just have to think a bit more carefully about how to explain/prove that this is an improvement, while also hoping that it won't break a lot of tests.