@tjstienstra 2024-03-26
It just removed the yaw frame in the formulation of the Whipple bicycle following Moore's convention. This causes a reduction in the number of operations in the EoMs after CSE from almost ~2500 to ~1900 (the number of operations before CSE only slightly increased from 230224 to 232020). This change seems a bit odd to me, as orient_body_fixed just does two rotations and stores them as one. It would be nice to investigate what calculation exactly profits from this change. If I have to guess then I would start with investigating vector differentiation.